Refactor botengine file & create events dir * Move all client event callbacks from botengine to events folder * Leave registerBotCommands implementation untouched as it requires a larger refactor later
